requirements/prod.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.5
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.5
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.16
django-cors-headers==3.8.0
django-environ==0.9.0
djangorestframework==3.12.4
docopt==0.6.2
drf-yasg==1.20.0
geographiclib==1.52
geopy==2.2.0
gTTS==2.2.3
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
imgkit==1.2.2
importlib-metadata==4.8.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.9.1
mailjet-rest==1.3.4
MarkupSafe==2.0.1
msgpack==1.0.4
msgpack-python==0.5.6
packaging==21.0
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
python-magic==0.4.24
pyttsx3==2.90
pytz==2021.1
redis==2.10.6
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.0.0
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.6
vtt-to-srt3==0.1.8.9
webvtt-py==0.4.6
wrapt==1.14.1
youtube-dl==2021.6.6
youtube-transcript-api==0.4.2
zipp==3.6.0
zope.interface==5.4.0